pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vm-sensor"
version = "0.1.0"
description = "Tkinter workstation scaffold for realtime capture and YOLO segmentation."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "numpy>=1.26.0",
  "opencv-python>=4.10.0",
  "Pillow>=10.4.0",
]

[project.optional-dependencies]
yolo = ["ultralytics>=8.3.0"]
basler = ["pypylon>=3.0.0"]
motor = ["pyserial>=3.5"]
dev = ["pytest>=8.3.0"]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
